+++ wallettemplate/src/main/java/wallettemplate/WalletPasswordController.java
@@ -30,6 +30,7 @@ import javafx.scene.control.ProgressIndicator;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.HBox;
+import org.bitcoinj.walletfx.application.WalletApplication;
 import org.bitcoinj.walletfx.overlay.OverlayController;
 import org.bitcoinj.walletfx.overlay.OverlayableStackPaneController;
 import org.slf4j.Logger;
@@ -79,13 +80,13 @@ public class WalletPasswordController implements OverlayController<WalletPasswor
             return;
         }
 
-        final KeyCrypterScrypt keyCrypter = (KeyCrypterScrypt) WalletTemplate.bitcoin.wallet().getKeyCrypter();
+        final KeyCrypterScrypt keyCrypter = (KeyCrypterScrypt) WalletApplication.bitcoin.wallet().getKeyCrypter();
         checkNotNull(keyCrypter);   // We should never arrive at this GUI if the wallet isn't actually encrypted.
         KeyDerivationTasks tasks = new KeyDerivationTasks(keyCrypter, password, getTargetTime()) {
             @Override
             protected final void onFinish(KeyParameter aesKey, int timeTakenMsec) {
                 checkGuiThread();
-                if (WalletTemplate.bitcoin.wallet().checkAESKey(aesKey)) {
+                if (WalletApplication.bitcoin.wallet().checkAESKey(aesKey)) {
                     WalletPasswordController.this.aesKey.set(aesKey);
                 } else {
                     log.warn("User entered incorrect password");
@@ -122,11 +123,11 @@ public class WalletPasswordController implements OverlayController<WalletPasswor
     // Writes the given time to the wallet as a tag so we can find it again in this class.
     public static void setTargetTime(Duration targetTime) {
         ByteString bytes = ByteString.copyFrom(Longs.toByteArray(targetTime.toMillis()));
-        WalletTemplate.bitcoin.wallet().setTag(TAG, bytes);
+        WalletApplication.bitcoin.wallet().setTag(TAG, bytes);
     }
 
     // Reads target time or throws if not set yet (should never happen).
     public static Duration getTargetTime() throws IllegalArgumentException {
-        return Duration.ofMillis(Longs.fromByteArray(WalletTemplate.bitcoin.wallet().getTag(TAG).toByteArray()));
+        return Duration.ofMillis(Longs.fromByteArray(WalletApplication.bitcoin.wallet().getTag(TAG).toByteArray()));
     }
 }
