+++ wallettemplate/src/main/java/wallettemplate/WalletTemplate.java
@@ -16,142 +16,33 @@
 
 package wallettemplate;
 
-import com.google.common.util.concurrent.Service;
-import javafx.application.Platform;
 import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.input.KeyCombination;
 import javafx.scene.layout.Pane;
-import javafx.stage.Stage;
 import org.bitcoinj.core.NetworkParameters;
-import org.bitcoinj.core.Utils;
-import org.bitcoinj.kits.WalletAppKit;
-import org.bitcoinj.params.RegTestParams;
 import org.bitcoinj.script.Script;
-import org.bitcoinj.utils.AppDataDirectory;
-import org.bitcoinj.utils.BriefLogFormatter;
-import org.bitcoinj.utils.Threading;
-import org.bitcoinj.wallet.DeterministicSeed;
-import org.bitcoinj.walletfx.application.AppDelegate;
-import org.bitcoinj.walletfx.utils.GuiUtils;
+import org.bitcoinj.walletfx.application.WalletApplication;
 
-import javax.annotation.Nullable;
-import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 
-import static org.bitcoinj.walletfx.utils.GuiUtils.informationalAlert;
-
 /**
- *
+ * Template implementation of WalletApplication
  */
-public class WalletTemplate implements AppDelegate {
-    static WalletAppKit bitcoin;
-    static WalletTemplate instance;
-
-    public final String applicationName;
-    private final String walletFileName;
-    public final NetworkParameters params;
-    public final Script.ScriptType preferredOutputScriptType;
-
-    private MainController controller;
+public class WalletTemplate extends WalletApplication {
 
     public WalletTemplate(String applicationName, NetworkParameters params, Script.ScriptType preferredOutputScriptType) {
-        instance = this;
-        this.applicationName = applicationName;
-        this.params = params;
-        this.preferredOutputScriptType = preferredOutputScriptType;
-        this.walletFileName = applicationName.replaceAll("[^a-zA-Z0-9.-]", "_") + "-" + params.getPaymentProtocolId();
+        super(applicationName, params, preferredOutputScriptType);
     }
 
     @Override
-    public void start(Stage mainWindow) throws Exception {
-        try {
-            realStart(mainWindow);
-        } catch (Throwable e) {
-            GuiUtils.crashAlert(e);
-            throw e;
-        }
-    }
-
-    private void realStart(Stage mainWindow) throws IOException {
-        instance = this;
-        // Show the crash dialog for any exceptions that we don't handle and that hit the main loop.
-        GuiUtils.handleCrashesOnThisThread();
-
-        if (Utils.isMac()) {
-            // We could match the Mac Aqua style here, except that (a) Modena doesn't look that bad, and (b)
-            // the date picker widget is kinda broken in AquaFx and I can't be bothered fixing it.
-            // AquaFx.style();
-        }
-
+    protected MainController loadController() throws IOException {
         // Load the GUI. The MainController class will be automagically created and wired up.
         URL location = getClass().getResource("main.fxml");
         FXMLLoader loader = new FXMLLoader(location);
         Pane mainUI = loader.load();
-        controller = loader.getController();
+        MainController controller = loader.getController();
 
-        Scene scene = controller.controllerStart(mainUI, "wallet.css");
-        mainWindow.setScene(scene);
-
-        // Make log output concise.
-        BriefLogFormatter.init();
-        // Tell bitcoinj to run event handlers on the JavaFX UI thread. This keeps things simple and means
-        // we cannot forget to switch threads when adding event handlers. Unfortunately, the DownloadListener
-        // we give to the app kit is currently an exception and runs on a library thread. It'll get fixed in
-        // a future version.
-        Threading.USER_THREAD = Platform::runLater;
-        // Create the app kit. It won't do any heavyweight initialization until after we start it.
-        setupWalletKit(null);
-
-        if (bitcoin.isChainFileLocked()) {
-            informationalAlert("Already running", "This application is already running and cannot be started twice.");
-            Platform.exit();
-            return;
-        }
-
-        mainWindow.show();
-
-        WalletSetPasswordController.estimateKeyDerivationTimeMsec();
-
-        bitcoin.addListener(new Service.Listener() {
-            @Override
-            public void failed(Service.State from, Throwable failure) {
-                GuiUtils.crashAlert(failure);
-            }
-        }, Platform::runLater);
-        bitcoin.startAsync();
-
-        scene.getAccelerators().put(KeyCombination.valueOf("Shortcut+F"), () -> bitcoin.peerGroup().getDownloadPeer().close());
-    }
-
-    public void setupWalletKit(@Nullable DeterministicSeed seed) {
-        // If seed is non-null it means we are restoring from backup.
-        File appDataDirectory = AppDataDirectory.get(applicationName).toFile();
-        bitcoin = new WalletAppKit(params, preferredOutputScriptType, null, appDataDirectory, walletFileName) {
-            @Override
-            protected void onSetupCompleted() {
-                Platform.runLater(controller::onBitcoinSetup);
-            }
-        };
-        // Now configure and start the appkit. This will take a second or two - we could show a temporary splash screen
-        // or progress widget to keep the user engaged whilst we initialise, but we don't.
-        if (params == RegTestParams.get()) {
-            bitcoin.connectToLocalHost();   // You should run a regtest mode bitcoind locally.
-        }
-        bitcoin.setDownloadListener(controller.progressBarUpdater())
-                .setBlockingStartup(false)
-                .setUserAgent(applicationName, "1.0");
-        if (seed != null)
-            bitcoin.restoreWalletFromSeed(seed);
-    }
-
-
-    @Override
-    public void stop() throws Exception {
-        bitcoin.stopAsync();
-        bitcoin.awaitTerminated();
-        // Forcibly terminate the JVM because Orchid likes to spew non-daemon threads everywhere.
-        Runtime.getRuntime().exit(0);
+        controller.controllerStart(mainUI, "wallet.css");
+        return controller;
     }
 }
